[linker] Extract suitable mobile tests to valid linker tests
