--- mcs/tools/linker/Tests/LinkerDescriptor/link-sdk.xml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<linker>
-	<assembly fullname="link-sdk">
-		<!-- TODO: these can be removed once PreserveAttribute works -->
-		<type fullname="LinkSdk.Aot.AotBugsTest" />
-		<type fullname="LinkSdk.AsyncTests" />
-		<type fullname="LinkSdk.Calendars.CalendarTest" />
-		<type fullname="LinkSdk.CanaryTest" />
-		<type fullname="LinkSdk.CrashExample" />
-		<type fullname="LinkSdk.CryptoTest" />
-		<type fullname="LinkSdk.DllImportTest" />
-		<type fullname="LinkSdk.LinkExtraDefsTest" />
-		<type fullname="LinkSdk.LinkSdkRegressionTest" />
-		<type fullname="LinkSdk.LocaleTest" />
-		<type fullname="LinkSdk.Net.Http.HttpClientHandlerTest" />
-		<type fullname="LinkSdk.Net.Http.HttpClientTest" />
-		<type fullname="LinkSdk.PclTest" />
-		<type fullname="LinkSdk.Serialization.XmlSerializationTest" />
-		<type fullname="LinkSdk.TaskBugsTest" />
-		<type fullname="Linker.Shared.Reflection.ReflectionTest" />
-	</assembly>
-	<!-- see LinkExtraDefsTest unit tests -->
-	<assembly fullname="mscorlib">
-		<type fullname="System.Security.PermissionSet" />
-	</assembly>
-	<assembly fullname="System">
-		<type fullname="System.Net.Mime.ContentType" >
-			<method name="ParseValue" />
-		</type>
-	</assembly>
-	<!-- end LinkExtraDefsTest preserves -->
-	<assembly fullname="System.Configuration">
-		<type fullname="System.Configuration.ExeConfigurationHost" preserve="methods" />
-	</assembly>
-	<assembly fullname="System">
-		<type fullname="System.Diagnostics.SystemDiagnosticsSection" preserve="methods" />
-		<type fullname="System.Diagnostics.FilterElement" preserve="methods" />
-		<type fullname="System.Net.Configuration.DefaultProxySection" preserve="methods" />
-		<type fullname="System.Net.Configuration.NetSectionGroup" preserve="methods" />
-		<type fullname="System.Net.Configuration.SettingsSection" preserve="methods" />
-		<type fullname="System.Net.Configuration.WebRequestModulesSection" preserve="methods" />
-		<type fullname="System.Net.HttpRequestCreator" preserve="methods" />
-		<type fullname="System.Net.FileWebRequestCreator" preserve="methods" />
-		<type fullname="System.Net.FtpWebRequestCreator" preserve="methods" />
-		<type fullname="System.ComponentModel.ArrayConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.BooleanConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.CharConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.CollectionConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.ComponentConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.CultureInfoConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.DateTimeOffsetConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.DecimalConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.DoubleConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.SByteConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.GuidConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.EnumConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.StringConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.SingleConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.TimeSpanConverter" preserve="methods" />
-		<type fullname="System.ComponentModel.Int16Converter" preserve="methods" />
-		<type fullname="System.ComponentModel.Int32Converter" preserve="methods" />
-		<type fullname="System.ComponentModel.Int64Converter" preserve="methods" />
-		<type fullname="System.ComponentModel.UInt16Converter" preserve="methods" />
-		<type fullname="System.ComponentModel.UInt32Converter" preserve="methods" />
-		<type fullname="System.ComponentModel.UInt64Converter" preserve="methods" />
-		<type fullname="System.ComponentModel.NullableConverter" preserve="methods" />
-		<type fullname="System.Net.Configuration.BypassElementCollection" preserve="methods" />
-		<type fullname="System.Net.Configuration.HttpWebRequestElement" preserve="methods" />
-		<type fullname="System.Net.Configuration.Ipv6Element" preserve="methods" />
-		<type fullname="System.Net.Configuration.PerformanceCountersElement" preserve="methods" />
-		<type fullname="System.Net.Configuration.ProxyElement" preserve="methods" />
-		<type fullname="System.Net.Configuration.ServicePointManagerElement" preserve="methods" />
-		<type fullname="System.Net.Configuration.SocketElement" preserve="methods" />
-		<type fullname="System.Net.Configuration.WebProxyScriptElement" preserve="methods" />
-		<type fullname="System.Net.Configuration.WebRequestModuleElementCollection" preserve="methods" />
-		<type fullname="System.Net.Configuration.WebRequestModuleElement" preserve="methods" />
-		<type fullname="System.UriTypeConverter" preserve="methods" />
-		<type fullname="System.Net.Configuration.ConnectionManagementSection" preserve="methods" />
-		<type fullname="System.Net.Configuration.ConnectionManagementElementCollection" preserve="methods" />
-	</assembly>
-	<assembly fullname="System.Runtime.Serialization">
-		<type fullname="System.Runtime.Serialization.Configuration.SerializationSectionGroup" preserve="methods"/>
-		<type fullname="System.Runtime.Serialization.Configuration.DataContractSerializerSection" preserve="methods" />
-		<type fullname="System.Runtime.Serialization.Configuration.DeclaredTypeElementCollection" preserve="methods" />
-	</assembly>
-</linker>
