--- mcs/tools/linker/Tests/link-sdk/Driver.cs
@@ -1,14 +1,0 @@
-using System;
-using System.IO;
-using NUnitLite.Runner;
-
-public class Driver
-{
-	public static int Main (string[] args)
-	{
-		var runner = new TextUI ();
-		runner.Execute (args);
-		
-		return (runner.Failure ? 1 : 0);
-	}
-}
