+++ mcs/tools/linker/Tests/mscorlib/test-calendar-01.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Globalization;
+
+public class C
+{
+	public static unsafe int Main ()
+	{
+		var ci = CultureInfo.GetCultureInfo ("ar");
+
+		// FIXME:
+//		if (ci.Calendar.ToString () != "System.Globalization.UmAlQuraCalendar")
+//			return 1;
+
+		return 0;
+	}
+}
