+++ mcs/tools/linker/Tests/mscorlib/test-crypto-01.cs
@@ -1,0 +1,13 @@
+using System.Security.Cryptography;
+
+class C
+{
+	public static int Main ()
+	{
+		Aes aes = Aes.Create ();
+		if (aes == null)
+			return 1;
+
+		return 0;
+	}
+}
